// frontend/src/api/generated/model/transportConfigOperationOperation.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type TransportConfigOperationOperation = typeof TransportConfigOperationOperation[keyof typeof TransportConfigOperationOperation];


export const TransportConfigOperationOperation = {
  create: 'create',
  update: 'update',
  delete: 'delete',
} as const;